+++ addon/components/basic-dropdown/content.js
@@ -127,8 +127,8 @@ export default Component.extend({
     // The following condition checks whether we need to open the dropdown - either because it was
     // closed and is now open or because it was open and then it was closed and opened pretty much at
     // the same time, indicated by `top`, `left` and `right` being null.
-    let {top, left, right} = this.getProperties('top', 'left', 'right');
-    if ((!oldDropdown.isOpen || (top === null && left === null && right === null)) && dropdown.isOpen) {
+    let {top, left, right, renderInPlace} = this.getProperties('top', 'left', 'right', 'renderInPlace');
+    if ((!oldDropdown.isOpen || (top === null && left === null && right === null && renderInPlace === false)) && dropdown.isOpen) {
       scheduleOnce('afterRender', this, this.open);
     } else if (oldDropdown.isOpen && !dropdown.isOpen) {
       this.close();